Feature/validator dlib check (#39)
* remove generate on the go training

this is just way too inefficient

* move load hair fast gan to utils

* add inference pipeline

* add inference pipeline

* separate align face responsibility from hair swap

* fix wrong module import

* update landmark detector downloader

update inference pipeline

* add dlib checking for validator

* move validation and insertion to follow patching policy

no db caching will be implemented and new tasks insertions will be added
using patches

* add tools for data integration

add db insert params
add db cache retrieval
add validate tasks params
add validate patch mode

* update newest cache file link

---------

+++ libs/face_parsing/models/utils.py
@@ -24,7 +24,7 @@ def get_mask_by_idx(
         logits = model(img)[0]      # 1 x C x H x W
         pred = logits.argmax(dim=1)  # 1 x H x W
 
-        mask = (pred == class_idx).to(torch.uint8)  # 1 x H x W
+        mask = (pred == class_idx).float()  # 1 x H x W
         masks.append(mask)
 
     return torch.stack(masks, dim=0)  # B x 1 x H x W
